popular 修正

+++ node/twitter-api/api/addPopularToDB.js
@@ -35,12 +35,19 @@ module.exports = async function addPopularToDB(token_type,arr){
                 if(flag) {
                     //tableない場合 -> 作成
                     await DB.queryp(`create table twitter_trend_popular${trend_id} (ids varchar(256))`);
+                    //リストに記録(更新)
+                    await DB.queryp(`insert into twitter_trend_popular${trend_id} (ids) values("${id_string}")`);
                 } else {
-                    //tableある場合 -> 削除
-                    await DB.queryp(`delete from twitter_trend_popular${trend_id}`);
+                    //tableある場合
+                    if(id_string != ""){
+                        //popularありの場合
+
+                        //前のデータ削除
+                        await DB.queryp(`delete from twitter_trend_popular${trend_id}`);
+                        //リストに記録(更新)
+                        await DB.queryp(`insert into twitter_trend_popular${trend_id} (ids) values("${id_string}")`);
+                    }
                 }
-                //リストに記録(更新)
-                await DB.queryp(`insert into twitter_trend_popular${trend_id} (ids) values("${id_string}")`);
 
                 resolve();
             } catch(e){
